directory restructure. distinguish plugins / extensions. begin extensions api
