+++ package.json
@@ -5,14 +5,22 @@
   "main": "index.js",
   "scripts": {
     "clean": "rimraf lib dist coverage",
-    "check": "npm run lint && npm run test",
     "start": "babel-node server/devServer.js",
+
+    "check": "npm run lint && npm run test",
+
     "lint": "eslint src server extensions; exit 0;",
     "lint-jscs": "jscs src",
+
     "test": "rm -rf ./storage/test && NODE_ENV=test mocha --recursive --compilers js:babel-register,css:test/css-null-compiler.js --require ./test/setup.js --timeout=10000",
     "test:watch": "npm test -- --watch",
     "nightwatch": "./node_modules/nightwatch/bin/nightwatch",
-    "postinstall": "./node_modules/selenium-standalone/bin/selenium-standalone install; node extensions/autogenRegistry.js > src/extensions/registry.js",
+    "selenium" : "./node_modules/selenium-standalone/bin/selenium-standalone install",
+
+    "postinstall": "npm run selenium",
+
+    "extensions" : "node extensions/autogenRegistry.js > src/extensions/registry.js",
+
     "start-docker": "docker build -t genome-designer . && docker run --privileged -v /var/run/docker.sock:/run/docker.sock -i -p 3000:3000 -t genome-designer",
     "create-ecoli-db": "python extensions/foundry/ecoli/generate_db.py jdon extensions/foundry/ecoli/ecoli.genes.json extensions/foundry/ecoli.json"
   },
