+++ plugins/compute/genbank_to_block/inputs/headers
@@ -1,0 +1,1 @@
+{"sessionkey":"4ea9a3c3263c566d3f5ea8b2eb11a81d6fdf8ba9","host":"http://0.0.0.0:3000"}
