+++ plugins/compute/genbank_to_block/outputs/block
@@ -1,0 +1,1 @@
+{"sequence": {"annotations": [{"isORF": false, "row": 0, "text": "gene", "start": 68, "textColor": "black", "strand": 1, "end": 4919, "color": "white", "id": "cc8e1b9e-c4b6-4bfe-a828-ad3290213e93"}, {"isORF": false, "row": 0, "text": "183K protein; position 3417..3419 is a readthrough codon", "start": 68, "textColor": "black", "strand": 1, "end": 4919, "color": "white", "id": "d1f9e428-49f5-4e44-92b7-665d11599978"}, {"isORF": false, "row": 0, "text": "gene", "start": 68, "textColor": "black", "strand": 1, "end": 3419, "color": "white", "id": "02171bdc-ff6b-4091-89a0-c3452e2f16de"}, {"isORF": false, "row": 0, "text": "126K protein", "start": 68, "textColor": "black", "strand": 1, "end": 3419, "color": "white", "id": "1bceee99-a386-4740-8266-befdb56c822e"}, {"isORF": false, "row": 0, "text": "gene", "start": 3494, "textColor": "black", "strand": 1, "end": 4919, "color": "white", "id": "48010ae6-80db-4c1a-88d5-83010c3f7ad7"}, {"isORF": false, "row": 0, "text": "54 kDa protein", "start": 3494, "textColor": "black", "strand": 1, "end": 4919, "color": "white", "id": "4629abc1-71dd-43c3-8cd9-cbea04a46664"}, {"isORF": false, "row": 0, "text": "gene", "start": 4902, "textColor": "black", "strand": 1, "end": 5709, "color": "white", "id": "d8f79339-4e71-430d-b84c-c02761fa38e8"}, {"isORF": false, "row": 0, "text": "CDS", "start": 4902, "textColor": "black", "strand": 1, "end": 5709, "color": "white", "id": "a9724ead-d955-43e4-97da-5aeb23e92e8f"}, {"isORF": false, "row": 0, "text": "gene", "start": 5665, "textColor": "black", "strand": 1, "end": 5788, "color": "white", "id": "32c1ed58-49dc-4977-9943-b2809bd24a33"}, {"isORF": false, "row": 0, "text": "CDS", "start": 5665, "textColor": "black", "strand": 1, "end": 5788, "color": "white", "id": "c60fca66-63e5-48e2-b21c-d99239aa8b6a"}, {"isORF": false, "row": 0, "text": "gene", "start": 5711, "textColor": "black", "strand": 1, "end": 6191, "color": "white", "id": "99270084-4adb-4f04-9a08-70db2a4d1efd"}, {"isORF": false, "row": 0, "text": "CDS", "start": 5711, "textColor": "black", "strand": 1, "end": 6191, "color": "white", "id": "ca711927-10ae-4f3b-a80e-9a0bb5047fcf"}], "url": null}, "rules": {}, "options": [], "notes": {}, "metadata": {"version": "0.0.0", "authors": [], "tags": {}}, "components": []}
