+++ plugins/compute/genbank_to_block/outputs/sequence
@@ -1,0 +1,1 @@
+GTATTTTTACAACAATTACCAACAACAACAAACAACAAACAACATTACAATTACTATTTACAATTACAATGGCATACACACAGACAGCTACCACATCAGCTTTGCTGGACACTGTCCGAGGAAACAACTCCTTGGTCAATGATCTAGCAAAGCGTCGTCTTTACGACACAGCGGTTGAAGAGTTTAACGCTCGTGACCGCAGGCCCAAGGTGAACTTTTCAAAAGTAATAAGCGAGGAGCAGACGCTTATTGCTACCCGGGCGTATCCAGAATTCCAAATTACATTTTATAACACGCAAAATGCCGTGCATTCGCTTGCAGGTGGATTGCGATCTTTAGAACTGGAATATCTGATGATGCAAATTCCCTACGGATCATTGACTTATGACATAGGCGGGAATTTTGCATCGCATCTGTTCAAGGGACGAGCATATGTACACTGCTGCATGCCCAACCTGGACGTTCGAGACATCATGCGGCACGAAGGCCAGAAAGACAGTATTGAACTATACCTTTCTAGGCTAGAGAGAGGGGGGAAAACAGTCCCCAACTTCCAAAAGGAAGCATTTGACAGATACGCAGAAATTCCTGAAGACGCTGTCTGTCACAATACTTTCCAGACAATGCGACATCAGCCGATGCAGCAATCAGGCAGAGTGTATGCCATTGCGCTACACAGCATATATGACATACCAGCCGATGAGTTCGGGGCGGCACTCTTGAGGAAAAATGTCCATACGTGCTATGCCGCTTTCCACTTCTCTGAGAACCTGCTTCTTGAAGATTCATACGTCAATTTGGACGAAATCAACGCGTGTTTTTCGCGCGATGGAGACAAGTTGACCTTTTCTTTTGCATCAGAGAGTACTCTTAATTATTGTCATAGTTATTCTAATATTCTTAAGTATGTGTGCAAAACTTACTTCCCGGCCTCTAATAGAGAGGTTTACATGAAGGAGTTTTTAGTCACCAGAGTTAATACCTGGTTTTGTAAGTTTTCTAGAATAGATACTTTTCTTTTGTACAAAGGTGTGGCCCATAAAAGTGTAGATAGTGAGCAGTTTTATACTGCAATGGAAGACGCATGGCATTACAAAAAGACTCTTGCAATGTGCAACAGCGAGAGAATCCTCCTTGAGGATTCATCATCAGTCAATTACTGGTTTCCCAAAATGAGGGATATGGTCATCGTACCATTATTCGACATTTCTTTGGAGACTAGTAAGAGGACGCGCAAGGAAGTCTTAGTGTCCAAGGATTTCGTGTTTACAGTGCTTAACCACATTCGAACATACCAGGCGAAAGCTCTTACATACGCAAATGTTTTGTCCTTTGTCGAATCGATTCGATCGAGGGTAATCATTAACGGTGTGACAGCGAGGTCCGAATGGGATGTGGACAAATCTTTGTTACAATCCTTGTCCATGACGTTTTACCTGCATACTAAGCTTGCCGTTCTAAAGGATGACTTACTGATTAGCAAGTTTAGTCTCGGTTCGAAAACGGTGTGCCAGCATGTGTGGGATGAGATTTCGCTGGCGTTTGGGAACGCATTTCCCTCCGTGAAAGAGAGGCTCTTGAACAGGAAACTTATCAGAGTGGCAGGCGACGCATTAGAGATCAGGGTGCCTGATCTATATGTGACCTTCCACGACAGATTAGTGACTGAGTACAAGGCCTCTGTGGACATGCCTGCGCTTGACATTAGGAAGAAGATGGAAGAAACGGAAGTGATGTACAATGCACTTTCAGAGTTATCGGTGTTAAGGGAGTCTGACAAATTCGATGTTGATGTTTTTTCCCAGATGTGCCAATCTTTGGAAGTTGACCCAATGACGGCAGCGAAGGTTATAGTCGCGGTCATGAGCAATGAGAGCGGTCTGACTCTCACATTTGAACGACCTACTGAGGCGAATGTTGCGCTAGCTTTACAGGATCAAGAGAAGGCTTCAGAAGGTGCTTTGGTAGTTACCTCAAGAGAAGTTGAAGAACCGTCCATGAAGGGTTCGATGGCCAGAGGAGAGTTACAATTAGCTGGTCTTGCTGGAGATCATCCGGAGTCGTCCTATTCTAAGAACGAGGAGATAGAGTCTTTAGAGCAGTTTCATATGGCAACGGCAGATTCGTTAATTCGTAAGCAGATGAGCTCGATTGTGTACACGGGTCCGATTAAAGTTCAGCAAATGAAAAACTTTATCGATAGCCTGGTAGCATCACTATCTGCTGCGGTGTCGAATCTCGTCAAGATCCTCAAAGATACAGCTGCTATTGACCTTGAAACCCGTCAAAAGTTTGGAGTCTTGGATGTTGCATCTAGGAAGTGGTTAATCAAACCAACGGCCAAGAGTCATGCATGGGGTGTTGTTGAAACCCACGCGAGGAAGTATCATGTGGCGCTTTTGGAATATGATGAGCAGGGTGTGGTGACATGCGATGATTGGAGAAGAGTAGCTGTCAGCTCTGAGTCTGTTGTTTATTCCGACATGGCGAAACTCAGAACTCTGCGCAGACTGCTTCGAAACGGAGAACCGCATGTCAGTAGCGCAAAGGTTGTTCTTGTGGACGGAGTTCCGGGCTGTGGGAAAACCAAAGAAATTCTTTCCAGGGTTAATTTTGATGAAGATCTAATTTTAGTACCTGGGAAGCAAGCCGCGGAAATGATCAGAAGACGTGCGAATTCCTCAGGGATTATTGTGGCCACGAAGGACAACGTTAAAACCGTTGATTCTTTCATGATGAATTTTGGGAAAAGCACACGCTGTCAGTTCAAGAGGTTATTCATTGATGAAGGGTTGATGTTGCATACTGGTTGTGTTAATTTTCTTGTGGCGATGTCATTGTGCGAAATTGCATATGTTTACGGAGACACACAGCAGATTCCATACATCAATAGAGTTTCAGGATTCCCGTACCCCGCCCATTTTGCCAAATTGGAAGTTGACGAGGTGGAGACACGCAGAACTACTCTCCGTTGTCCAGCCGATGTCACACATTATCTGAACAGGAGATATGAGGGCTTTGTCATGAGCACTTCTTCGGTTAAAAAGTCTGTTTCGCAGGAGATGGTCGGCGGAGCCGCCGTGATCAATCCGATCTCAAAACCCTTGCATGGCAAGATCCTGACTTTTACCCAATCGGATAAAGAAGCTCTGCTTTCAAGAGGGTATTCAGATGTTCACACTGTGCATGAAGTGCAAGGCGAGACATACTCTGATGTTTCACTAGTTAGGTTAACCCCTACACCAGTCTCCATCATTGCAGGAGACAGCCCACATGTTTTGGTCGCATTGTCAAGGCACACCTGTTCGCTCAAGTACTACACTGTTGTTATGGATCCTTTAGTTAGTATCATTAGAGATCTAGAGAAACTTAGCTCGTACTTGTTAGATATGTATAAGGTCGATGCAGGAACACAATAGCAATTACAGATTGACTCGGTGTTCAAAGGTTCCAATCTTTTTGTTGCAGCGCCAAAGACTGGTGATATTTCTGATATGCAGTTTTACTATGATAAGTGTCTCCCAGGCAACAGCACCATGATGAATAATTTTGATGCTGTTACCATGAGGTTGACTGACATTTCATTGAATGTCAAAGATTGCATATTGGATATGTCTAAGTCTGTTGCTGCGCCTAAGGATCAAATCAAACCACTAATACCTATGGTACGAACGGCGGCAGAAATGCCACGCCAGACTGGACTATTGGAAAATTTAGTGGCGATGATTAAAAGGAACTTTAACGCACCCGAGTTGTCTGGCATCATTGATATTGAAAATACTGCATCTTTAGTTGTAGATAAGTTTTTTGATAGTTATTTGCTTAAAGAAAAAAGAAAACCAAATAAAAATGTTTCTTTGTTCAGTAGAGAGTCTCTCAATAGATGGTTAGAAAAGCAGGAACAGGTAACAATAGGCCAGCTCGCAGATTTTGATTTTGTAGATTTGCCAGCAGTTGATCAGTACAGACACATGATTAAAGCACAACCCAAGCAAAAATTGGACACTTCAATCCAAACGGAGTACCCGGCTTTGCAGACGATTGTGTACCATTCAAAAAAGATCAATGCAATATTTGGCCCGTTGTTTAGTGAGCTTACTAGGCAATTACTGGACAGTGTTGATTCGAGCAGATTTTTGTTTTTCACAAGAAAGACACCAGCGCAGATTGAGGATTTCTTCGGAGATCTCGACAGTCATGTGCCGATGGATGTCTTGGAGCTGGATATATCAAAATACGACAAATCTCAGAATGAATTCCACTGTGCAGTAGAATACGAGATCTGGCGAAGATTGGGTTTTGAAGACTTCTTGGGAGAAGTTTGGAAACAAGGGCATAGAAAGACCACCCTCAAGGATTATACCGCAGGTATAAAAACTTGCATCTGGTATCAAAGAAAGAGCGGGGACGTCACGACGTTCATTGGAAACACTGTGATCATTGCTGCATGTTTGGCCTCGATGCTTCCGATGGAGAAAATAATCAAAGGAGCCTTTTGCGGTGACGATAGTCTGCTGTACTTTCCAAAGGGTTGTGAGTTTCCGGATGTGCAACACTCCGCGAATCTTATGTGGAATTTTGAAGCAAAACTGTTTAAAAAACAGTATGGATACTTTTGCGGAAGATATGTAATACATCACGACAGAGGATGCATTGTGTATTACGATCCCCTAAAGTTGATCTCGAAACTTGGTGCTAAACACATCAAGGATTGGGAACACTTGGAGGAGTTCAGAAGGTCTCTTTGTGATGTTGCTGTTTCGTTGAACAATTGTGCGTATTACACACAGTTGGACGACGCTGTATGGGAGGTTCATAAGACCGCCCCTCCAGGTTCGTTTGTTTATAAAAGTCTGGTGAAGTATTTGTCTGATAAAGTTCTTTTTAGAAGTTTGTTTATAGATGGCTCTAGTTGTTAAAGGAAAAGTGAATATCAATGAGTTTATCGACCTGACAAAAATGGAGAAGATCTTACCGTCGATGTTTACCCCTGTAAAGAGTGTTATGTGTTCCAAAGTTGATAAAATAATGGTTCATGAGAATGAGTCATTGTCAGAGGTGAACCTTCTTAAAGGAGTTAAGCTTATTGATAGTGGATACGTCTGTTTAGCCGGTTTGGTCGTCACGGGCGAGTGGAACTTGCCTGACAATTGCAGAGGAGGTGTGAGCGTGTGTCTGGTGGACAAAAGGATGGAAAGAGCCGACGAGGCCACTCTCGGATCTTACTACACAGCAGCTGCAAAGAAAAGATTTCAGTTCAAGGTCGTTCCCAATTATGCTATAACCACCCAGGACGCGATGAAAAACGTCTGGCAAGTTTTAGTTAATATTAGAAATGTGAAGATGTCAGCGGGTTTCTGTCCGCTTTCTCTGGAGTTTGTGTCGGTGTGTATTGTTTATAGAAATAATATAAAATTAGGTTTGAGAGAGAAGATTACAAACGTGAGAGACGGAGGGCCCATGGAACTTACAGAAGAAGTCGTTGATGAGTTCATGGAAGATGTCCCTATGTCGATCAGGCTTGCAAAGTTTCGATCTCGAACCGGAAAAAAGAGTGATGTCCGCAAAGGGAAAAATAGTAGTAATGATCGGTCAGTGCCGAACAAGAACTATAGAAATGTTAAGGATTTTGGAGGAATGAGTTTTAAAAAGAATAATTTAATCGATGATGATTCGGAGGCTACTGTCGCCGAATCGGATTCGTTTTAAATATGTCTTACAGTATCACTACTCCATCTCAGTTCGTGTTCTTGTCATCAGCGTGGGCCGACCCAATAGAGTTAATTAATTTATGTACTAATGCCTTAGGAAATCAGTTTCAAACACAACAAGCTCGAACTGTCGTTCAAAGACAATTCAGTGAGGTGTGGAAACCTTCACCACAAGTAACTGTTAGGTTCCCTGACAGTGACTTTAAGGTGTACAGGTACAATGCGGTATTAGACCCGCTAGTCACAGCACTGTTAGGTGCATTCGACACTAGAAATAGAATAATAGAAGTTGAAAATCAGGCGAACCCCACGACTGCCGAAACGTTAGATGCTACTCGTAGAGTAGACGACGCAACGGTGGCCATAAGGAGCGCGATAAATAATTTAATAGTAGAATTGATCAGAGGAACCGGATCTTATAATCGGAGCTCTTTCGAGAGCTCTTCTGGTTTGGTTTGGACCTCTGGTCCTGCAACTTGAGGTAGTCAAGATGCATAATAAATAACGGATTGTGTCCGTAATCACACGTGGTGCGTACGATAACGCATAGTGTTTTTCCCTCCACTTAAATCGAAGGGTTGTGTCTTGGATCGCGCGGGTCAAATGTATATGGTTCATATACATCCGCAGGCACGTAATAAAGCGAGGGGTTCGAATCCCCCCGTTACCCCCGGTAGGGGCCCA
