+++ plugins/compute/translate_dna_example/Dockerfile
