+++ plugins/compute/translate_dna_example/inputs/DNA
@@ -1,0 +1,1 @@
+AACTTGTCCACTGTA
