+++ plugins/compute/translate_dna_example/inputs/headers
@@ -1,0 +1,1 @@
+{}
