+++ plugins/compute/translate_dna_example/translate.py
