+++ plugins/compute/translate_dna_example/workflow.yaml
