+++ plugins/search/nucleotide/index.js
