+++ server/devServer.js
@@ -6,11 +6,12 @@ import config from './../webpack.config.dev.js';
 
 import dataRouter from './data/index';
 import fileRouter from './file/index';
-import extensionsRouter from '../extensions/index';
-import extRouter from '../extensions/compute/api';
-import importRouter from '../extensions/convert/import';
-import exportRouter from '../extensions/convert/export';
-import searchRouter from '../extensions/search/search';
+import extensionsRouter from './extensions/index';
+
+import computeRouter from '../plugins/compute/api';
+import importRouter from '../plugins/convert/import';
+import exportRouter from '../plugins/convert/export';
+import searchRouter from '../plugins/search/search';
 
 import { authRouter } from './utils/authentication';
 
@@ -24,32 +25,26 @@ const compiler = webpack(config);
 //logging middleware
 app.use(morgan('dev'));
 
-// Register Hotloading Middleware
-// ----------------------------------------------------
-
+//Hotloading Middleware
 app.use(require('webpack-dev-middleware')(compiler, {
   noInfo: true,
   publicPath: config.output.publicPath,
 }));
-
 app.use(require('webpack-hot-middleware')(compiler));
 
-
-// Register API middleware
-// ----------------------------------------------------
-
+//authentication
 app.use('/auth', authRouter);
 
-// all these should require authentication middleware
-
+//primary routes
 app.use('/data', dataRouter);
 app.use('/file', fileRouter);
+app.use('/extensions', extensionsRouter);
 
-app.use('/compute', extRouter);
+//hardwired extensions
+app.use('/compute', computeRouter);
 app.use('/import', importRouter);
 app.use('/export', exportRouter);
 app.use('/search', searchRouter);
-app.use('/extensions', extensionsRouter);
 
 // Register Client Requests, delegate routing to client
 // ----------------------------------------------------
